Move response example to a file
Preparing for possibly adding additional tests using different responses later
